+++ bin/blueprint.php
@@ -1,34 +1,39 @@
 <?php
 
-// Parses the data from a YAML format
-$parser = new Rougin\Blueprint\Parser(
-    new Symfony\Component\Yaml\Parser
-);
+// Includes the Composer Autoloader
+require realpath('vendor') . '/autoload.php';
 
-$content = $parser->parse('blueprint.yml', getcwd());
-
-$blueprint = new Rougin\Blueprint\Blueprint(
-    new Symfony\Component\Console\Application,
-    new Auryn\Injector
-);
-
-// Information of the command application
-$blueprint->console->setName('Blueprint');
-$blueprint->console->setVersion('0.1.5');
-
-// Adds a "init" command if the file does not exists
-if ( ! file_exists('blueprint.yml')) {
-    $blueprint->console->add(
-        new Rougin\Blueprint\Commands\InitializationCommand
-    );
-
-    return $blueprint;
+if ( ! defined('BLUEPRINT_FILENAME')) {
+    define('BLUEPRINT_FILENAME', 'blueprint.yml');
 }
 
+// Creates a new instance of Blueprint class
+$injector = new Auryn\Injector;
+$console = new Symfony\Component\Console\Application;
+$app = new Rougin\Blueprint\Blueprint($console, $injector);
+
+// Information of the application
+$app->console->setName('Blueprint');
+$app->console->setVersion('0.2.0');
+
+// Adds a "init" command if the file does not exists
+if ( ! file_exists(BLUEPRINT_FILENAME)) {
+    $command = new Rougin\Blueprint\Commands\InitializationCommand;
+
+    $app->console->add($command);
+
+    return $app->run();
+}
+
+// Parses the data from a YAML format
+$file = file_get_contents(BLUEPRINT_FILENAME);
+$file = str_replace('%%CURRENT_DIRECTORY%%', getcwd(), $file);
+$content = Symfony\Component\Yaml\Yaml::parse($file);
+
 // Set paths from the parsed result
-$blueprint
-    ->setTemplatePath($content['paths']['templates'])
+$app->setTemplatePath($content['paths']['templates'])
     ->setCommandPath($content['paths']['commands'])
     ->setCommandNamespace($content['namespaces']['commands']);
 
-return $blueprint;
+// Run the application
+$app->run();