+++ tests/TestApp/Commands/AbstractCommand.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace Rougin\Blueprint\TestApp\Commands;
+
+use Rougin\Blueprint\Commands\AbstractCommand as BaseCommand;
+
+/**
+ * Abstract Command
+ * 
+ * @package TestApp
+ * @author  Rougin Royce Gutib <rougingutib@gmail.com>
+ */
+abstract class AbstractCommand extends BaseCommand {}