+++ CHANGELOG.md
@@ -2,6 +2,18 @@
 
 All Notable changes to `Blueprint` will be documented in this file
 
+## [0.2.0](https://github.com/rougin/blueprint/compare/v0.1.5...v0.2.0)
+
+### Added
+- Unit tests for `Blueprint`
+- `Common\File` class for handling files
+
+### Changed
+- Moved preloading of `Twig_Environment` to `Blueprint::setTemplatePath`
+
+### Removed
+- `Parser` class
+
 ## [0.1.5](https://github.com/rougin/blueprint/compare/v0.1.4...v0.1.5) - 2016-02-02
 
 ### Changed